backend/requirements.txt
fastapi
uvicorn
langchain
langchain-google-genai
pypdf
python-docx
pillow
streamlit
python-dotenv
chromadb
reportlab
langchain_experimental
qtpy